requirements.txt
openai>=1.0.0
langgraph>=0.2.0
serpapi>=0.1.5
beautifulsoup4>=4.12.0
requests>=2.31.0
flask>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.0
selenium>=4.15.0
webdriver-manager>=4.0.0
mcp>=0.1.0
aiohttp>=3.9.0
